refactor(consumer) : 패키지 구조 설정
config : 설정 관련
dto : dto들
service : application 레벨에서의 처리
repository : DB 접근 레이어
