+++ spring-consumer/src/main/java/org/consumer/repository/DelayedTransferRepository.java
@@ -1,0 +1,4 @@
+package org.consumer.repository;
+
+public class DelayedTransferRepository {
+}
