+++ spring-consumer/src/main/java/org/consumer/service/Consumer.java
@@ -1,7 +1,8 @@
-package org.consumer;
+package org.consumer.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.consumer.dto.MessageDto;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
@@ -12,9 +13,9 @@ public class Consumer {
     private final static String GROUP_ID = "group1";
 
     @KafkaListener(topics = TOPIC, groupId = GROUP_ID)
-    public void receive(ConsumerRecord<String, DelayedTransferDto> record) {
+    public void receive(ConsumerRecord<String, MessageDto> record) {
         final String userId = record.key();
-        final DelayedTransferDto message = record.value();
+        final MessageDto message = record.value();
         log.info("Consumed message`s key: {}", userId);
         log.info("Consumed message`s key: {}", message.toString());
     }
